+++ src/UniTask/Assets/Plugins/UniTask/Runtime/AsyncReactiveProperty.cs
@@ -1,5 +1,4 @@
-﻿using Cysharp.Threading.Tasks.Linq;
-using System;
+﻿using System;
 using System.Threading;
 
 namespace Cysharp.Threading.Tasks
