+++ src/UniTask/Assets/Plugins/UniTask/Runtime/Linq/UniTask.Linq.asmdef
@@ -1,0 +1,15 @@
+{
+    "name": "UniTask.Linq",
+    "references": [
+        "UniTask"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
