+++ src/UniTask/Assets/Plugins/UniTask/Runtime/MoveNextSource.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dc4c5dc2a5f246e4f8df44cab735826c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
