+++ src/UniTask/Assets/Plugins/UniTask/Runtime/Triggers/AsyncTriggerBase.cs
@@ -1,6 +1,5 @@
 ﻿#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 
-using Cysharp.Threading.Tasks.Linq;
 using System;
 using System.Threading;
 using UnityEngine;
