+++ src/UniTask/Assets/Plugins/UniTask/Runtime/_InternalVisibleTo.cs
@@ -1,0 +1,3 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("UniTask.Linq")]
