+++ src/UniTask/Assets/Plugins/UniTask/Runtime/_InternalVisibleTo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8507e97eb606fad4b99c6edf92e19cb8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
